+++ src/Calculator.java
@@ -1,14 +1,14 @@
 public class Calculator {
-    public int addNumbers(int num1, int num2) {
+    public double addNumbers(double num1, double num2) {
         return num1 + num2;
     }
-    public int subtractNumbers(int num1, int num2) {
+    public double subtractNumbers(double num1, double num2) {
         return num1 - num2;
     }
-    public int multiplyNumbers(int num1, int num2) {
+    public double multiplyNumbers(double num1, double num2) {
         return num1 * num2;
     }
-    public int divideNumbers(int num1, int num2) {
+    public double divideNumbers(double num1, double num2) {
         return num1 / num2;
     }
 }
